Move parallel fixtures to common_parallel.py
